Support tracking with a "top-level" function.
To do this, only specify a single function in the WORKLOAD env variable.
Then the tracer will track all functions called by that function and its
children.  Note that this means a single function in the WORKLOAD is
treated as top-level, even if the user does not intend for this to be
the case, because in practice they should be the same thing (if they're
not, the user is probably doing something wrong).

One major difference is that in top-level mode, we instrument ALL
instructions and dynamically determine whether to include that
instruction in the trace or not. This is to avoid having to rewrite
the opt pass as a ModulePass and construct a function call tree (which
is the proper way to do this). Maybe one day we'll get around to
implementing it like that...

This also simplifies tracking of C++ templated functions. To do so, just
provide the root stem of the function name, without template arguments.

Confirmed to produce the same trace as before on aes, with and without
top-level mode enabled.

Change-Id: I3fb906432de4b21aa7f1addfe1116729f93c23e1