ast-pass: Introduce an AST pass to get label information.
Labeled statements will make it infinitely easier to specify loop
unrolling parameters (which currently relies on line numbers), but label
names are not available in emitted IR. However, they are available in
the AST. This pass walks the AST and produces a file mapping labels to
line numbers, which the LLVM-Tracer can read and use when producing a
trace.

A small testcase is provided.

BUG=ALADDIN-63

Change-Id: I7001dca55d2dec442954c94abbfe17b312378f72