full_trace: Prepend the contents of labelmap to the dynamic trace.
The tracer reads the contents of the labelmap file, if present, and
embeds the contents into the instrumented binary. A callback is inserted
in the main() body of the binary to write the preserved labelmap file
into the dynamic trace. Then the labelmap file is deleted, thus keeping
it strictly a temporary file. Since the instrumented binary contains
the labelmap file contents, it remains completely standalone.

BUG=ALADDIN-63

Change-Id: Iaef713186ebe5b354e08a6c1e243384a4c22585a