Enable multiple 'top-level' functions.
If the command line arg -trace-all-callees is given to the opt pass,
then all callees of every function specified in WORKLOAD will be traced.
This differs from old behavior, where if multiple functions were
specified in WORKLOAD, only those functions would be traced (and not
unspecified callees).

By default, -trace-all-callees is set to false to preserve the old
behavior.

Change-Id: I525faa20e483f5af1aab308b981ce128d0333dce