Fix missing nodes in multiple invocations bug.
If a top level function was called more than once, the first node of the
second (and third and all subsequent calls) would not be logged, due to
an error in the logic.

The logic that determines whether to log the current instruction or not
has now been formalized into a truth table (since it was getting pretty
hairy otherwise). There are three possible results:

  - LOG_AND_CONTINUE: log this instruction and the next ones.
  - LOG_AND_STOP: log this instruction but not the next ones.
  - DO_NOT_LOG: do not log this instruction.

Change-Id: I76c77d1884bdd1850e218a72317102fb7b5ed789