For bitcasts of pointers w/o named registers, use the input operand name.
Suppose we have a bitcast like this:

  %1 = bitcast i32* %array to [16 x i32]*, !dbg !125

This would generate trace output like this:

  0,65,kernel,0,1,44,0
  1,64,8726336,1,array,
  r,64,8726336,1,1,

If we use the casted pointer, we've lost the name "array", even though
the underlying address is still the same. This creates problems when we
pass the casted pointer to a function, since the forwarded name will be
"1" rather than "array". So, for all bitcasts of pointers that do not
have a named output register, rename the output operand to be the input
operand. This will produce tracer output like so:

  0,65,kernel,0,1,44,0
  1,64,8726336,1,array,
  r,64,8726336,1,array,

Change-Id: Idcaf64d5fb33946e7f66e3459ce4e20f0479cd26