Add loop hierarchy information to the trace.
This is accomplished by appending the loop depth of a basic block to its
ID with a colon. Any change in the loop depth will indicate a transition
to a different loop nesting level. Disjoint loops, even those at the
same depth, will have the different basic block names (which precede the
colon).

Change-Id: Ibe01a691c439d5d6dfe7645e2b88e3a66ba60ecc