Basic support for modeling fixed-point op performance.
This change allows a user to mark all floating point operations inside a
function as fixed point, by suffixing the function name with "_fxp". If
this is done, then the trace will be printed in the same way, except
that relevant FP opcodes will be substituted with integer opcodes. The
VALUES printed will still be in floating-point format. This means that
all values are 100% consistent with the original FP implementation and
frees us from having to deal with low-precision correctness problems.
This patch is only for modeling the PERFORMANCE of fixed-point
operations.

Note that this patch also does not specify the bitwidth itself. A power
modeling tool should handle bitwidths separately, possibly with an
additional file that maps functions to their corresponding bitwidth, or
a fine-grained parsing of the execution trace to mark specific ops with
specific bitwidths.

Change-Id: Ic2afaf9760ad06049f6d9510f31abd4d3d88f7e9