Add support for printing the values of vector operands.
Vector operands are nontrivial to print because they are larger than 64
bits, and many variable sizes means they cannot just be passed by value
to the tracing functions. The most general way is to use a pointer to
the first byte of the data, but we cannot be guaranteed that an
arbitrary vector operand already has memory address. The solution is to
allocate a local array big enough to store the value, load the value
into that array, and give the pointer of that local array to the
tracing function.

We only support GCC vector literals (no support for NEON or OpenCL
vector extensions).

Example vector syntax:

  // Create a 8-wide vector of 32-bit signed integers.
  typedef int v8si __attribute__((__vector_size__(32)));
  v8si vector_value = (v8si){0,1,2,3,4,5,6,7};
  v8si vector_array[4];

Change-Id: I47b5ff5f94e72900883565c4177d463911c4d713