Print the forwarded function call argument names *after* entry.
Rather than inserting the instrumentation calls before the Call
instruction, we split it up into two parts: the caller part gets
inserted right before the Call as before, but the callee part gets
inserted after entering that function.

This resolves the cross-module function call problems with the tracer.
When a function is not defined in the module (i.e. *.c source file) it
is called from, the names of the function arguments are unavailable at
the point of the Call instruction, which defeats the purpose of printing
the "forwarded" callee arg names (currently, they get printed as empty
strings).  The only way in the past to fix this was to force
LLVM-Tracer to do full-program analysis by concatenating all source
files into a single large file. Now, we can compile different modules
separately, link, and we'll still get all the function argument names!

This does change the trace format:
  1. Forwarded argument lines all come after ALL caller argument lines,
     rather than interleaved.
  2. Pointer arguments are printed as hex with a "0x" prefix.

Example:

  0,12,test,0:0,0:0-4,49,8
  3,64,0x402aa0,1,called_function,
  1,64,0x7ffc4ccc00e0,1,array0,
  2,64,0x7ffc4ccc00d0,1,array1,   <- 1st forward line moved from here...
  f,64,0x7ffc4ccc00e0,1,array0,   <- ...to here (after all caller args).
  f,64,0x7ffc4ccc00d0,1,array1,

Change-Id: I2d20c2f9844517bbc57bd7dacf3e1ece84da7b41