Upgrade LLVM-Tracer to v2.0, using LLVM 6.0.
This is a major update that brings LLVM-Tracer up to date with LLVM at
last. Many LLVM APIs have changed over the course of 3 major versions.
Now, we will no longer require users to build LLVM from source (for the
foreseeable future). LLVM 6 binaries, libraries, and dev headers are
available for most platforms in pre-built installation packages, and
this is the recommended way to go.

This update was driven by the need to support compiling with C++ code.
To be clear: LLVM-Tracer does not support instrumentation of C++ code or
any code that relies on C++ features, and the previous commit that added
support to skip functions with C++ linkage would have been sufficient.
However, we needed support for ABI tags (introduced in GCC 5, with support
added in LLVM 3.9), among other changes. Rather than upgrade to only
LLVM 3.9 (still many years old), it made more sense to just jump
straight to LLVM 6.

Change-Id: Ic739dca971b6bbf1f1435fb026ff065b69e7a72e