Build a cache of debug info from the current function.
This allows us to use generated debug information for names of values,
rather than relying on a Value to have a name (Value->hasName() ==
true). This has two benefits:

  1) Release builds of LLVM may omit the value names when emitting LLVM
     IR, and there is no way to control this behavior from the clang
     driver (instead, you'd need to modify the cc1 command to restore
     value names). So debug info is the only way to get value names
     back.
  2) Debug info is more likely to be directly relatable to the source,
     and in the dynamic trace, we don't care about SSA form anyways.

This patch scans the function for any debug intrinsics and caches all
information pertaining to the names of Values before adding
instrumentation.

Change-Id: I3652509a6fbfe53e5980b1403e1a24794d5cf694