Fix bugs with trace state management when entering/exiting top level functions.
When exiting a top level function, the tracer deletes the thread-local
trace pointer. Thus, if another top level function is called, upon
entry, it must also reinitialize the trace state.

This also fixes a segfault in which if fin_main is called, the trace
pointer would be null (it was deleted by fin_toplevel) and then deleted.

Change-Id: I105d1c35e0aaa683926902460e5655879f306f2a