Add a special setSamplingFactor instruction.
We will use this special instruction to set the sampling factor in
the accelerator. The setSamplingFactor takes two arguments, the first
is the name of the loop that is intended to be sampled, and the second
is the sampling factor. An example shows below:

  int top_level(int a, int b, int size, int sample) {
    setSamplingFactor("loop", size / sample);
    loop:
    for (int i = 0; i < sample; i++)
      a += i;
    return a + b;
  }

  int main() {
    int c = top_level(1, 2, 10, 1);
    printf("Result = %d\n", c);
    return 0;
  }

This "loop" is sampled, with the sampling factor 10. The generated
node in the trace is:

  0,13,top_level,3:0,5,105,1
  3,64,0x402ba0,1,setSamplingFactor,
  1,64,loop,0, ,
  2,32,10,1,4,
  f,64,0x552470,1,loop_name,
  f,32,10,1,factor,
  r,32,0,1,5,

Change-Id: I03856b9eaf81e05b449b06d27f668d5843abaec7