Simplify the AST pass's handling of C++ function declarations.
Trying to determine the "type of language" a function was written in is
ambiguous and cannot be determined purely from linkage or storage
classes. Instead, visit all function declarations except for C++ member
function declarations and function templates, which is a single
unambiguous check. When printing the labelmap, print the fully qualified
name (which includes namespaces).

This also adds a test.cpp file with some of these test cases; however,
we don't have a test runner yet so the test need to be run manually:

    get-labeled-stmts test.c -- -I/path/to/clang/6/headers
    get-labeled-stmts test.cpp -- -I/path/to/clang/6/headers

Change-Id: I24ed67ca6638bb3518da59df3bae50588e877a61